Added auth middleware
In the middleware we check if their token in the request, parsed by cookie-parser library from HTTPS request header to cookies property in the req object. If their is token we verified it and  add user key with user data from the decoded token as value in the req object. Also we add IsAuthenticated property in the req object with true as value. If the token is invalid we redirect the user to the logic page. If their no token in the req.cookies that means the user is Guest.